+++ Sources/Sentry/SentryTracer.m
@@ -16,17 +16,23 @@
 static const void *spanTimestampObserver = &spanTimestampObserver;
 
 /**
- * The maximum amount of seconds the app start measurment end time and the start time of the
+ * The maximum amount of seconds the app start measurement end time and the start time of the
  * transaction are allowed to be apart.
  */
 static const NSTimeInterval SENTRY_APP_START_MEASUREMENT_DIFFERENCE = 5.0;
 
+@interface
+SentryTracer ()
+
+@property (nonatomic, strong) SentrySpan *rootSpan;
+@property (nonatomic, strong) NSMutableArray<id<SentrySpan>> *children;
+@property (nonatomic, strong) SentryHub *hub;
+@property (nonatomic) SentrySpanStatus finishStatus;
+@property (nonatomic) BOOL isWaitingForChildren;
+
+@end
+
 @implementation SentryTracer {
-    SentrySpan *_rootSpan;
-    NSMutableArray<id<SentrySpan>> *_children;
-    SentryHub *_hub;
-    SentrySpanStatus _finishStatus;
-    BOOL _isFinished;
     BOOL _waitForChildren;
 
 #if SENTRY_HAS_UIKIT
@@ -49,13 +55,13 @@ static const NSTimeInterval SENTRY_APP_START_MEASUREMENT_DIFFERENCE = 5.0;
                            waitForChildren:(BOOL)waitForChildren
 {
     if ([super init]) {
-        _rootSpan = [[SentrySpan alloc] initWithTracer:self context:transactionContext];
+        self.rootSpan = [[SentrySpan alloc] initWithTracer:self context:transactionContext];
         self.name = transactionContext.name;
-        _children = [[NSMutableArray alloc] init];
-        _hub = hub;
-        _isFinished = YES;
+        self.children = [[NSMutableArray alloc] init];
+        self.hub = hub;
+        self.isWaitingForChildren = NO;
         _waitForChildren = waitForChildren;
-        _finishStatus = kSentrySpanStatusUndefined;
+        self.finishStatus = kSentrySpanStatusUndefined;
 
 #if SENTRY_HAS_UIKIT
         _startTimeChanged = NO;
@@ -104,11 +110,11 @@ static const NSTimeInterval SENTRY_APP_START_MEASUREMENT_DIFFERENCE = 5.0;
         [child addObserver:self
                 forKeyPath:NSStringFromSelector(@selector(timestamp))
                    options:NSKeyValueObservingOptionNew
-                   context:&spanTimestampObserver];
+                   context:nil];
     }
 
-    @synchronized(_children) {
-        [_children addObject:child];
+    @synchronized(self.children) {
+        [self.children addObject:child];
     }
 
     return child;
@@ -122,13 +128,12 @@ static const NSTimeInterval SENTRY_APP_START_MEASUREMENT_DIFFERENCE = 5.0;
                         change:(NSDictionary<NSKeyValueChangeKey, id> *)change
                        context:(void *)context
 {
-    if (context == spanTimestampObserver &&
-        [keyPath isEqualToString:NSStringFromSelector(@selector(timestamp))]) {
+    if ([keyPath isEqualToString:NSStringFromSelector(@selector(timestamp))]) {
         SentrySpan *finishedSpan = object;
         if (finishedSpan.timestamp != nil) {
             [finishedSpan removeObserver:self
                               forKeyPath:NSStringFromSelector(@selector(timestamp))
-                                 context:&spanTimestampObserver];
+                                 context:nil];
             [self canBeFinished];
         }
     }
@@ -136,27 +141,27 @@ static const NSTimeInterval SENTRY_APP_START_MEASUREMENT_DIFFERENCE = 5.0;
 
 - (SentrySpanContext *)context
 {
-    return _rootSpan.context;
+    return self.rootSpan.context;
 }
 
 - (NSDate *)timestamp
 {
-    return _rootSpan.timestamp;
+    return self.rootSpan.timestamp;
 }
 
 - (void)setTimestamp:(NSDate *)timestamp
 {
-    _rootSpan.timestamp = timestamp;
+    self.rootSpan.timestamp = timestamp;
 }
 
 - (NSDate *)startTimestamp
 {
-    return _rootSpan.startTimestamp;
+    return self.rootSpan.startTimestamp;
 }
 
 - (void)setStartTimestamp:(NSDate *)startTimestamp
 {
-    _rootSpan.startTimestamp = startTimestamp;
+    self.rootSpan.startTimestamp = startTimestamp;
 
 #if SENTRY_HAS_UIKIT
     _startTimeChanged = YES;
@@ -165,17 +170,17 @@ static const NSTimeInterval SENTRY_APP_START_MEASUREMENT_DIFFERENCE = 5.0;
 
 - (NSDictionary<NSString *, id> *)data
 {
-    return _rootSpan.data;
+    return self.rootSpan.data;
 }
 
 - (BOOL)isFinished
 {
-    return _rootSpan.isFinished;
+    return self.rootSpan.isFinished;
 }
 
 - (void)setDataValue:(nullable id)value forKey:(NSString *)key
 {
-    [_rootSpan setDataValue:value forKey:key];
+    [self.rootSpan setDataValue:value forKey:key];
 }
 
 - (void)finish
@@ -185,7 +190,7 @@ static const NSTimeInterval SENTRY_APP_START_MEASUREMENT_DIFFERENCE = 5.0;
 
 - (void)finishWithStatus:(SentrySpanStatus)status
 {
-    _isFinished = YES;
+    self.isWaitingForChildren = YES;
     _finishStatus = status;
     [self canBeFinished];
 }
@@ -203,7 +208,7 @@ static const NSTimeInterval SENTRY_APP_START_MEASUREMENT_DIFFERENCE = 5.0;
 
 - (void)canBeFinished
 {
-    if (!_isFinished || (_waitForChildren && [self hasUnfinishedChildren]))
+    if (!self.isWaitingForChildren || (_waitForChildren && [self hasUnfinishedChildren]))
         return;
 
     [_rootSpan finishWithStatus:_finishStatus];
@@ -226,7 +231,7 @@ static const NSTimeInterval SENTRY_APP_START_MEASUREMENT_DIFFERENCE = 5.0;
 
 - (SentryTransaction *)toTransaction
 {
-    SentryAppStartMeasurement *appStartMeasurement = [self getAppStartMeasurment];
+    SentryAppStartMeasurement *appStartMeasurement = [self getAppStartMeasurement];
 
     NSArray<id<SentrySpan>> *appStartSpans = [self buildAppStartSpans:appStartMeasurement];
 
@@ -253,43 +258,34 @@ static const NSTimeInterval SENTRY_APP_START_MEASUREMENT_DIFFERENCE = 5.0;
     return transaction;
 }
 
-- (SentryAppStartMeasurement *)getAppStartMeasurment
+- (nullable SentryAppStartMeasurement *)getAppStartMeasurement
 {
-    SentryAppStartMeasurement *appStartMeasurement = nil;
-
     // Only send app start measurement for transactions generated by auto performance
     // instrumentation.
     if (![self.context.operation isEqualToString:SENTRY_VIEWCONTROLLER_RENDERING_OPERATION]) {
-        return appStartMeasurement;
+        return nil;
     }
 
-    // Double-Checked Locking to avoid acquiring unnecessary locks.
-    if (SentrySDK.appStartMeasurement == nil) {
-        return appStartMeasurement;
+    SentryAppStartMeasurement *measurement = [SentrySDK getAndResetAppStartMeasurement];
+    if (measurement == nil) {
+        return nil;
     }
 
-    @synchronized(SentrySDK.appStartMeasurementLock) {
-        if (SentrySDK.appStartMeasurement != nil) {
-            NSDate *appStartTimestamp = SentrySDK.appStartMeasurement.appStartTimestamp;
-            NSDate *appStartEndTimestamp =
-                [appStartTimestamp dateByAddingTimeInterval:SentrySDK.appStartMeasurement.duration];
+    NSDate *appStartTimestamp = measurement.appStartTimestamp;
+    NSDate *appStartEndTimestamp =
+        [appStartTimestamp dateByAddingTimeInterval:measurement.duration];
 
-            NSTimeInterval difference =
-                [appStartEndTimestamp timeIntervalSinceDate:self.startTimestamp];
+    NSTimeInterval difference = [appStartEndTimestamp timeIntervalSinceDate:self.startTimestamp];
 
-            // If the difference between the end of the app start and the beginning of the current
-            // transaction is smaller than SENTRY_APP_START_MEASUREMENT_DIFFERENCE. With this we
-            // avoid messing up transactions too much.
-            if (difference <= SENTRY_APP_START_MEASUREMENT_DIFFERENCE
-                && difference >= -SENTRY_APP_START_MEASUREMENT_DIFFERENCE) {
-                appStartMeasurement = SentrySDK.appStartMeasurement;
-            }
-
-            SentrySDK.appStartMeasurement = nil;
-        }
+    // If the difference between the end of the app start and the beginning of the current
+    // transaction is smaller than SENTRY_APP_START_MEASUREMENT_DIFFERENCE. With this we
+    // avoid messing up transactions too much.
+    if (difference > SENTRY_APP_START_MEASUREMENT_DIFFERENCE
+        || difference < -SENTRY_APP_START_MEASUREMENT_DIFFERENCE) {
+        return nil;
     }
 
-    return appStartMeasurement;
+    return measurement;
 }
 
 - (NSArray<SentrySpan *> *)buildAppStartSpans: