+++ CHANGELOG.md
@@ -10,6 +10,8 @@ feat: Measure slow and frozen frames (#1123)
 - fix: Set description for performance tracker spans (#1042)
 - fix: Capturing Child Spans (#1139)
 - feat: Auto UI Performance Instrumentation (#1105)
+- feat: Measure app start time (#1111)
+- feat: Auto UI Performance Instrumentation (#1105, #1150, #1136, #1139, #1042)
 
 ## 7.1.3